rename: router.ts -> sidebar.ts로 이름 변경
